Propagate some additional callee argument attributes before inlining
To avoid losing information, we can propagate some access attribute
from the to-be-inlined callee to its callsites.

If the argument used the callsite is the same as the argument for the
to-be-inlined callee, we can often propagate it.

This patch adds propagation for `nonnull`, `deferereancable`,
`deferereancable_or_null`, and `align`.